feat: add crop image pipe to game card component
The crop image pipe is added to the game card component to allow for cropping of the background image. This pipe takes in the image URL and appends a crop parameter to it, which specifies the dimensions of the cropped image. This will ensure that the background image is displayed in a consistent size across all game cards.